Move fileutils to a standalone module.
