+++ app/build.gradle
@@ -197,6 +197,7 @@ dependencies {
     implementation project(':HttpRequest')
     implementation project(':httptask')
     implementation project(':urlutils')
+    implementation project(':fileutils')
     implementation project(':icon')
     implementation project(':logger')
     implementation project(':threadutils')
