+++ app/src/main/java/org/mozilla/focus/persistence/TabModelStore.java
@@ -6,9 +6,9 @@ import android.preference.PreferenceManager;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
+import org.mozilla.fileutils.FileUtils;
 import org.mozilla.focus.Inject;
 import org.mozilla.focus.R;
-import org.mozilla.focus.utils.FileUtils;
 import org.mozilla.rocket.tabs.Session;
 
 import java.io.File;
