+++ app/src/main/java/org/mozilla/focus/utils/StorageUtils.java
@@ -10,6 +10,7 @@ import android.os.Environment;
 import android.support.annotation.NonNull;
 import android.support.annotation.VisibleForTesting;
 
+import org.mozilla.fileutils.FileUtils;
 import org.mozilla.focus.R;
 
 import java.io.File;
