+++ components/utils/fileutils/src/main/java/org/mozilla/fileutils/SerializedItem.java
@@ -1,4 +1,4 @@
-package org.mozilla.focus.utils;
+package org.mozilla.fileutils;
 
 import java.io.Serializable;
 import java.nio.ByteBuffer;
