+++ settings.gradle
@@ -1,8 +1,9 @@
-include ':feature-tabs', ':HttpRequest', ':httptask', ':logger', ':threadutils', 'urlutils', ':icon'
+include ':feature-tabs', ':fileutils', ':HttpRequest', ':httptask', ':logger', ':threadutils', 'urlutils', ':icon'
 project(':feature-tabs').projectDir = new File(rootDir, 'components/feature/tabs')
 project(':HttpRequest').projectDir = new File(rootDir, 'components/utils/HttpRequest')
 project(':httptask').projectDir = new File(rootDir, 'components/utils/httptask')
 project(':urlutils').projectDir = new File(rootDir, 'components/utils/urlutils')
+project(':fileutils').projectDir = new File(rootDir, 'components/utils/fileutils')
 project(':icon').projectDir = new File(rootDir, 'components/utils/icon')
 project(':logger').projectDir = new File(rootDir, 'components/utils/logger')
 project(':threadutils').projectDir = new File(rootDir, 'components/utils/threadutils')
